+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Fixed aura icon tooltips not showing reliably in WoW Classic versions [Comment #8183, #8226].
 * Fixed "Force View for Target" not switching from Headline View to Healthbar View when targeting a unit that was not taking damage (e.g. friendly units out of combat) [Comment #8308].
+* Fixed Blizzard name font changes not applying reliably by re-applying nameplate font overrides (including SystemFont_NamePlate_Outlined) after delayed UI updates [Comment #8230].